:truck: moved api logic to main app and composable
